@@ -11,6 +11,7 @@ import { Packet } from "./packet";
 import type { GameType } from "./player";
 import { DisplaySlot, ObjectiveSortOrder, ScoreboardId } from "./scoreboard";
 import minecraft = require('../minecraft');
+import MinecraftPacketIds = minecraft.MinecraftPacketIds;
 
 /** @deprecated */
 @nativeClass(null)
@@ -1666,167 +1667,167 @@ export namespace NpcDialoguePacket {
 
 /** @deprecated */
 export const PacketIdToType = {
-    0x01: LoginPacket,
-    0x02: PlayStatusPacket,
-    0x03: ServerToClientHandshakePacket,
-    0x04: ClientToServerHandshakePacket,
-    0x05: DisconnectPacket,
-    0x06: ResourcePacksInfoPacket,
-    0x07: ResourcePackStackPacket,
-    0x08: ResourcePackClientResponsePacket,
-    0x09: TextPacket,
-    0x0a: SetTimePacket,
-    0x0b: StartGamePacket,
-    0x0c: AddPlayerPacket,
-    0x0d: AddActorPacket,
-    0x0e: RemoveActorPacket,
-    0x0f: AddItemActorPacket,
-    0x11: TakeItemActorPacket,
-    0x12: MoveActorAbsolutePacket,
-    0x13: MovePlayerPacket,
-    0x14: RiderJumpPacket,
-    0x15: UpdateBlockPacket,
-    0x16: AddPaintingPacket,
-    0x17: TickSyncPacket,
-    0x18: LevelSoundEventPacketV1,
-    0x19: LevelEventPacket,
-    0x1a: BlockEventPacket,
-    0x1b: ActorEventPacket,
-    0x1c: MobEffectPacket,
-    0x1d: UpdateAttributesPacket,
-    0x1e: InventoryTransactionPacket,
-    0x1f: MobEquipmentPacket,
-    0x20: MobArmorEquipmentPacket,
-    0x21: InteractPacket,
-    0x22: BlockPickRequestPacket,
-    0x23: ActorPickRequestPacket,
-    0x24: PlayerActionPacket,
-    0x26: HurtArmorPacket,
-    0x27: SetActorDataPacket,
-    0x28: SetActorMotionPacket,
-    0x29: SetActorLinkPacket,
-    0x2a: SetHealthPacket,
-    0x2b: SetSpawnPositionPacket,
-    0x2c: AnimatePacket,
-    0x2d: RespawnPacket,
-    0x2e: ContainerOpenPacket,
-    0x2f: ContainerClosePacket,
-    0x30: PlayerHotbarPacket,
-    0x31: InventoryContentPacket,
-    0x32: InventorySlotPacket,
-    0x33: ContainerSetDataPacket,
-    0x34: CraftingDataPacket,
-    0x35: CraftingEventPacket,
-    0x36: GuiDataPickItemPacket,
-    0x37: AdventureSettingsPacket,
-    0x38: BlockActorDataPacket,
-    0x39: PlayerInputPacket,
-    0x3a: LevelChunkPacket,
-    0x3b: SetCommandsEnabledPacket,
-    0x3c: SetDifficultyPacket,
-    0x3d: ChangeDimensionPacket,
-    0x3e: SetPlayerGameTypePacket,
-    0x3f: PlayerListPacket,
-    0x40: SimpleEventPacket,
-    0x41: TelemetryEventPacket,
-    0x42: SpawnExperienceOrbPacket,
-    0x43: MapItemDataPacket,
-    0x44: MapInfoRequestPacket,
-    0x45: RequestChunkRadiusPacket,
-    0x46: ChunkRadiusUpdatedPacket,
-    0x47: ItemFrameDropItemPacket,
-    0x48: GameRulesChangedPacket,
-    0x49: CameraPacket,
-    0x4a: BossEventPacket,
-    0x4b: ShowCreditsPacket,
-    0x4c: AvailableCommandsPacket,
-    0x4d: CommandRequestPacket,
-    0x4e: CommandBlockUpdatePacket,
-    0x4f: CommandOutputPacket,
-    0x52: ResourcePackDataInfoPacket,
-    0x53: ResourcePackChunkDataPacket,
-    0x54: ResourcePackChunkRequestPacket,
-    0x55: TransferPacket,
-    0x56: PlaySoundPacket,
-    0x57: StopSoundPacket,
-    0x58: SetTitlePacket,
-    0x59: AddBehaviorTreePacket,
-    0x5a: StructureBlockUpdatePacket,
-    0x5b: ShowStoreOfferPacket,
-    0x5c: PurchaseReceiptPacket,
-    0x5d: PlayerSkinPacket,
-    0x5e: SubClientLoginPacket,
-    0x5f: WSConnectPacket,
-    0x60: SetLastHurtByPacket,
-    0x61: BookEditPacket,
-    0x62: NpcRequestPacket,
-    0x63: PhotoTransferPacket,
-    0x64: ModalFormRequestPacket,
-    0x65: ModalFormResponsePacket,
-    0x66: ServerSettingsRequestPacket,
-    0x67: ServerSettingsResponsePacket,
-    0x68: ShowProfilePacket,
-    0x69: SetDefaultGameTypePacket,
-    0x6a: RemoveObjectivePacket,
-    0x6b: SetDisplayObjectivePacket,
-    0x6c: SetScorePacket,
-    0x6d: LabTablePacket,
-    0x6e: UpdateBlockPacketSynced,
-    0x6f: MoveActorDeltaPacket,
-    0x70: SetScoreboardIdentityPacket,
-    0x71: SetLocalPlayerAsInitializedPacket,
-    0x72: UpdateSoftEnumPacket,
-    0x73: NetworkStackLatencyPacket,
-    0x75: ScriptCustomEventPacket,
-    0x76: SpawnParticleEffectPacket,
-    0x77: AvailableActorIdentifiersPacket,
-    0x78: LevelSoundEventPacketV2,
-    0x79: NetworkChunkPublisherUpdatePacket,
-    0x7a: BiomeDefinitionList,
-    0x7b: LevelSoundEventPacket,
-    0x7c: LevelEventGenericPacket,
-    0x7d: LecternUpdatePacket,
-    0x80: RemoveEntityPacket,
-    0x81: ClientCacheStatusPacket,
-    0x82: OnScreenTextureAnimationPacket,
-    0x83: MapCreateLockedCopy,
-    0x84: StructureTemplateDataRequestPacket,
-    0x85: StructureTemplateDataExportPacket,
-    0x87: ClientCacheBlobStatusPacket,
-    0x88: ClientCacheMissResponsePacket,
-    0x89: EducationSettingsPacket,
-    0x8a: EmotePacket,
-    0x8b: MultiplayerSettingsPacket,
-    0x8c: SettingsCommandPacket,
-    0x8d: AnvilDamagePacket,
-    0x8e: CompletedUsingItemPacket,
-    0x8f: NetworkSettingsPacket,
-    0x90: PlayerAuthInputPacket,
-    0x91: CreativeContentPacket,
-    0x92: PlayerEnchantOptionsPacket,
-    0x93: ItemStackRequest,
-    0x94: ItemStackResponse,
-    0x95: PlayerArmorDamagePacket,
-    0x96: CodeBuilderPacket,
-    0x97: UpdatePlayerGameTypePacket,
-    0x98: EmoteListPacket,
-    0x99: PositionTrackingDBServerBroadcastPacket,
-    0x9a: PositionTrackingDBClientRequestPacket,
-    0x9b: DebugInfoPacket,
-    0x9c: PacketViolationWarningPacket,
-    0x9d: MotionPredictionHintsPacket,
-    0x9e: AnimateEntityPacket,
-    0x9f: CameraShakePacket,
-    0xa0: PlayerFogPacket,
-    0xa1: CorrectPlayerMovePredictionPacket,
-    0xa2: ItemComponentPacket,
-    0xa3: FilterTextPacket,
-    0xa4: ClientboundDebugRendererPacket,
-    0xa5: SyncActorPropertyPacket,
-    0xa6: AddVolumeEntityPacket,
-    0xa7: RemoveVolumeEntityPacket,
-    0xa8: SimulationTypePacket,
-    0xa9: NpcDialoguePacket,
+    [MinecraftPacketIds.Login]: LoginPacket,
+    [MinecraftPacketIds.PlayStatus]: PlayStatusPacket,
+    [MinecraftPacketIds.ServerToClientHandshake]: ServerToClientHandshakePacket,
+    [MinecraftPacketIds.ClientToServerHandshake]: ClientToServerHandshakePacket,
+    [MinecraftPacketIds.Disconnect]: DisconnectPacket,
+    [MinecraftPacketIds.ResourcePacksInfo]: ResourcePacksInfoPacket,
+    [MinecraftPacketIds.ResourcePackStack]: ResourcePackStackPacket,
+    [MinecraftPacketIds.ResourcePackClientResponse]: ResourcePackClientResponsePacket,
+    [MinecraftPacketIds.Text]: TextPacket,
+    [MinecraftPacketIds.SetTime]: SetTimePacket,
+    [MinecraftPacketIds.StartGame]: StartGamePacket,
+    [MinecraftPacketIds.AddPlayer]: AddPlayerPacket,
+    [MinecraftPacketIds.AddActor]: AddActorPacket,
+    [MinecraftPacketIds.RemoveActor]: RemoveActorPacket,
+    [MinecraftPacketIds.AddItemActor]: AddItemActorPacket,
+    [MinecraftPacketIds.TakeItemActor]: TakeItemActorPacket,
+    [MinecraftPacketIds.MoveActorAbsolute]: MoveActorAbsolutePacket,
+    [MinecraftPacketIds.MovePlayer]: MovePlayerPacket,
+    [MinecraftPacketIds.RiderJump]: RiderJumpPacket,
+    [MinecraftPacketIds.UpdateBlock]: UpdateBlockPacket,
+    [MinecraftPacketIds.AddPainting]: AddPaintingPacket,
+    [MinecraftPacketIds.TickSync]: TickSyncPacket,
+    [MinecraftPacketIds.LevelSoundEventPa]: LevelSoundEventPacketV1,
+    [MinecraftPacketIds.LevelEvent]: LevelEventPacket,
+    [MinecraftPacketIds.BlockEvent]: BlockEventPacket,
+    [MinecraftPacketIds.ActorEvent]: ActorEventPacket,
+    [MinecraftPacketIds.MobEffect]: MobEffectPacket,
+    [MinecraftPacketIds.UpdateAttributes]: UpdateAttributesPacket,
+    [MinecraftPacketIds.InventoryTransaction]: InventoryTransactionPacket,
+    [MinecraftPacketIds.MobEquipment]: MobEquipmentPacket,
+    [MinecraftPacketIds.MobArmorEquipment]: MobArmorEquipmentPacket,
+    [MinecraftPacketIds.Interact]: InteractPacket,
+    [MinecraftPacketIds.BlockPickRequest]: BlockPickRequestPacket,
+    [MinecraftPacketIds.ActorPickRequest]: ActorPickRequestPacket,
+    [MinecraftPacketIds.PlayerAction]: PlayerActionPacket,
+    [MinecraftPacketIds.HurtArmor]: HurtArmorPacket,
+    [MinecraftPacketIds.SetActorData]: SetActorDataPacket,
+    [MinecraftPacketIds.SetActorMotion]: SetActorMotionPacket,
+    [MinecraftPacketIds.SetActorLink]: SetActorLinkPacket,
+    [MinecraftPacketIds.SetHealth]: SetHealthPacket,
+    [MinecraftPacketIds.SetSpawnPosition]: SetSpawnPositionPacket,
+    [MinecraftPacketIds.Animate]: AnimatePacket,
+    [MinecraftPacketIds.Respawn]: RespawnPacket,
+    [MinecraftPacketIds.ContainerOpen]: ContainerOpenPacket,
+    [MinecraftPacketIds.ContainerClose]: ContainerClosePacket,
+    [MinecraftPacketIds.PlayerHotbar]: PlayerHotbarPacket,
+    [MinecraftPacketIds.InventoryContent]: InventoryContentPacket,
+    [MinecraftPacketIds.InventorySlot]: InventorySlotPacket,
+    [MinecraftPacketIds.ContainerSetData]: ContainerSetDataPacket,
+    [MinecraftPacketIds.CraftingData]: CraftingDataPacket,
+    [MinecraftPacketIds.CraftingEvent]: CraftingEventPacket,
+    [MinecraftPacketIds.GuiDataPickItem]: GuiDataPickItemPacket,
+    [MinecraftPacketIds.AdventureSettings]: AdventureSettingsPacket,
+    [MinecraftPacketIds.BlockActorData]: BlockActorDataPacket,
+    [MinecraftPacketIds.PlayerInput]: PlayerInputPacket,
+    [MinecraftPacketIds.LevelChunk]: LevelChunkPacket,
+    [MinecraftPacketIds.SetCommandsEnabled]: SetCommandsEnabledPacket,
+    [MinecraftPacketIds.SetDifficulty]: SetDifficultyPacket,
+    [MinecraftPacketIds.ChangeDimension]: ChangeDimensionPacket,
+    [MinecraftPacketIds.SetPlayerGameType]: SetPlayerGameTypePacket,
+    [MinecraftPacketIds.PlayerList]: PlayerListPacket,
+    [MinecraftPacketIds.SimpleEvent]: SimpleEventPacket,
+    [MinecraftPacketIds.Event]: TelemetryEventPacket,
+    [MinecraftPacketIds.SpawnExperienceOrb]: SpawnExperienceOrbPacket,
+    [MinecraftPacketIds.ClientboundMapItemData]: MapItemDataPacket,
+    [MinecraftPacketIds.MapInfoRequest]: MapInfoRequestPacket,
+    [MinecraftPacketIds.RequestChunkRadius]: RequestChunkRadiusPacket,
+    [MinecraftPacketIds.ChunkRadiusUpdated]: ChunkRadiusUpdatedPacket,
+    [MinecraftPacketIds.ItemFrameDropItem]: ItemFrameDropItemPacket,
+    [MinecraftPacketIds.GameRulesChanged]: GameRulesChangedPacket,
+    [MinecraftPacketIds.Camera]: CameraPacket,
+    [MinecraftPacketIds.BossEvent]: BossEventPacket,
+    [MinecraftPacketIds.ShowCredits]: ShowCreditsPacket,
+    [MinecraftPacketIds.AvailableCommands]: AvailableCommandsPacket,
+    [MinecraftPacketIds.CommandRequest]: CommandRequestPacket,
+    [MinecraftPacketIds.CommandBlockUpdate]: CommandBlockUpdatePacket,
+    [MinecraftPacketIds.CommandOutput]: CommandOutputPacket,
+    [MinecraftPacketIds.ResourcePackDataInfo]: ResourcePackDataInfoPacket,
+    [MinecraftPacketIds.ResourcePackChunkData]: ResourcePackChunkDataPacket,
+    [MinecraftPacketIds.ResourcePackChunkRequest]: ResourcePackChunkRequestPacket,
+    [MinecraftPacketIds.Transfer]: TransferPacket,
+    [MinecraftPacketIds.PlaySound]: PlaySoundPacket,
+    [MinecraftPacketIds.StopSound]: StopSoundPacket,
+    [MinecraftPacketIds.SetTitle]: SetTitlePacket,
+    [MinecraftPacketIds.AddBehaviorTree]: AddBehaviorTreePacket,
+    [MinecraftPacketIds.StructureBlockUpdate]: StructureBlockUpdatePacket,
+    [MinecraftPacketIds.ShowStoreOffer]: ShowStoreOfferPacket,
+    [MinecraftPacketIds.PurchaseReceipt]: PurchaseReceiptPacket,
+    [MinecraftPacketIds.PlayerSkin]: PlayerSkinPacket,
+    [MinecraftPacketIds.SubClientLogin]: SubClientLoginPacket,
+    [MinecraftPacketIds.AutomationClientConnect]: WSConnectPacket,
+    [MinecraftPacketIds.SetLastHurtBy]: SetLastHurtByPacket,
+    [MinecraftPacketIds.BookEdit]: BookEditPacket,
+    [MinecraftPacketIds.NpcRequest]: NpcRequestPacket,
+    [MinecraftPacketIds.PhotoTransfer]: PhotoTransferPacket,
+    [MinecraftPacketIds.ModalFormRequest]: ModalFormRequestPacket,
+    [MinecraftPacketIds.ModalFormResponse]: ModalFormResponsePacket,
+    [MinecraftPacketIds.ServerSettingsRequest]: ServerSettingsRequestPacket,
+    [MinecraftPacketIds.ServerSettingsResponse]: ServerSettingsResponsePacket,
+    [MinecraftPacketIds.ShowProfile]: ShowProfilePacket,
+    [MinecraftPacketIds.SetDefaultGameType]: SetDefaultGameTypePacket,
+    [MinecraftPacketIds.RemoveObjective]: RemoveObjectivePacket,
+    [MinecraftPacketIds.SetDisplayObjective]: SetDisplayObjectivePacket,
+    [MinecraftPacketIds.SetScore]: SetScorePacket,
+    [MinecraftPacketIds.LabTable]: LabTablePacket,
+    [MinecraftPacketIds.UpdateBlockPacket]: UpdateBlockPacketSynced,
+    [MinecraftPacketIds.MoveActorDelta]: MoveActorDeltaPacket,
+    [MinecraftPacketIds.SetScoreboardIdentity]: SetScoreboardIdentityPacket,
+    [MinecraftPacketIds.SetLocalPlayerAsInitialized]: SetLocalPlayerAsInitializedPacket,
+    [MinecraftPacketIds.UpdateSoftEnum]: UpdateSoftEnumPacket,
+    [MinecraftPacketIds.NetworkStackLatency]: NetworkStackLatencyPacket,
+    [MinecraftPacketIds.ScriptCustomEvent]: ScriptCustomEventPacket,
+    [MinecraftPacketIds.SpawnParticleEffect]: SpawnParticleEffectPacket,
+    [MinecraftPacketIds.AvailableActorIdentifiers]: AvailableActorIdentifiersPacket,
+    [MinecraftPacketIds.LevelSoundEventPa]: LevelSoundEventPacketV2,
+    [MinecraftPacketIds.NetworkChunkPublisherUpdate]: NetworkChunkPublisherUpdatePacket,
+    [MinecraftPacketIds.BiomeDefiniti]: BiomeDefinitionList,
+    [MinecraftPacketIds.LevelSoundEvent]: LevelSoundEventPacket,
+    [MinecraftPacketIds.LevelEventGeneric]: LevelEventGenericPacket,
+    [MinecraftPacketIds.LecternUpdate]: LecternUpdatePacket,
+    [MinecraftPacketIds.RemoveEntity]: RemoveEntityPacket,
+    [MinecraftPacketIds.ClientCacheStatus]: ClientCacheStatusPacket,
+    [MinecraftPacketIds.OnScreenTextureAnimation]: OnScreenTextureAnimationPacket,
+    [MinecraftPacketIds.MapCreateLock]: MapCreateLockedCopy,
+    [MinecraftPacketIds.StructureTemplateDataRequest]: StructureTemplateDataRequestPacket,
+    [MinecraftPacketIds.StructureTemplateDataResponse]: StructureTemplateDataExportPacket,
+    [MinecraftPacketIds.ClientCacheBlobStatus]: ClientCacheBlobStatusPacket,
+    [MinecraftPacketIds.ClientCacheMissResponse]: ClientCacheMissResponsePacket,
+    [MinecraftPacketIds.EducationSettings]: EducationSettingsPacket,
+    [MinecraftPacketIds.Emote]: EmotePacket,
+    [MinecraftPacketIds.MultiplayerSettings]: MultiplayerSettingsPacket,
+    [MinecraftPacketIds.SettingsCommand]: SettingsCommandPacket,
+    [MinecraftPacketIds.AnvilDamage]: AnvilDamagePacket,
+    [MinecraftPacketIds.CompletedUsingItem]: CompletedUsingItemPacket,
+    [MinecraftPacketIds.NetworkSettings]: NetworkSettingsPacket,
+    [MinecraftPacketIds.PlayerAuthInput]: PlayerAuthInputPacket,
+    [MinecraftPacketIds.CreativeContent]: CreativeContentPacket,
+    [MinecraftPacketIds.PlayerEnchantOptions]: PlayerEnchantOptionsPacket,
+    [MinecraftPacketIds.ItemStackR]: ItemStackRequest,
+    [MinecraftPacketIds.ItemStackRe]: ItemStackResponse,
+    [MinecraftPacketIds.PlayerArmorDamage]: PlayerArmorDamagePacket,
+    [MinecraftPacketIds.CodeBuilder]: CodeBuilderPacket,
+    [MinecraftPacketIds.UpdatePlayerGameType]: UpdatePlayerGameTypePacket,
+    [MinecraftPacketIds.EmoteList]: EmoteListPacket,
+    [MinecraftPacketIds.PositionTrackingDBServerBroadcast]: PositionTrackingDBServerBroadcastPacket,
+    [MinecraftPacketIds.PositionTrackingDBClientRequest]: PositionTrackingDBClientRequestPacket,
+    [MinecraftPacketIds.DebugInfo]: DebugInfoPacket,
+    [MinecraftPacketIds.PacketViolationWarning]: PacketViolationWarningPacket,
+    [MinecraftPacketIds.MotionPredictionHints]: MotionPredictionHintsPacket,
+    [MinecraftPacketIds.AnimateEntity]: AnimateEntityPacket,
+    [MinecraftPacketIds.CameraShake]: CameraShakePacket,
+    [MinecraftPacketIds.PlayerFog]: PlayerFogPacket,
+    [MinecraftPacketIds.CorrectPlayerMovePrediction]: CorrectPlayerMovePredictionPacket,
+    [MinecraftPacketIds.ItemComponent]: ItemComponentPacket,
+    [MinecraftPacketIds.FilterText]: FilterTextPacket,
+    [MinecraftPacketIds.ClientboundDebugRenderer]: ClientboundDebugRendererPacket,
+    [MinecraftPacketIds.SyncActorProperty]: SyncActorPropertyPacket,
+    [MinecraftPacketIds.AddVolumeEntity]: AddVolumeEntityPacket,
+    [MinecraftPacketIds.RemoveVolumeEntity]: RemoveVolumeEntityPacket,
+    [MinecraftPacketIds.SimulationType]: SimulationTypePacket,
+    [MinecraftPacketIds.NpcDialogue]: NpcDialoguePacket,
 };
 (PacketIdToType as any).__proto__ = null;
 /** @deprecated */
