@@ -2,16 +2,20 @@
  * @internal
  */
 
-const nullScope:UnusedName.Scope = {
-    existName() { return false; },
-    canAccessGlobalName() { return false; }
-};
+export interface ScopeMethod {
+    existName(name:string):boolean;
+}
+
+export namespace ScopeMethod {
+    export const empty:ScopeMethod = {
+        existName(name) {
+            return false;
+        }
+    };
+}
 
 export class UnusedName {
     private readonly generatedNames = new Map<string, number>();
-
-    constructor(public readonly scope:UnusedName.Scope = nullScope) {
-    }
 
     hasName(name:string):boolean {
         return this.generatedNames.has(name);
@@ -21,12 +25,12 @@ export class UnusedName {
         return this.generatedNames.delete(name);
     }
 
-    makeName(name:string):string {
+    makeName(scope:ScopeMethod, name:string):string {
         let counter:number|undefined;
         for (;;) {
             counter = this.generatedNames.get(name);
             if (counter == null) {
-                if (this.scope.existName(name)) {
+                if (scope.existName(name)) {
                     this.generatedNames.set(name, counter = 1);
                     break;
                 }
@@ -37,17 +41,10 @@ export class UnusedName {
         }
         for (;;) {
             const nname = name + (++counter);
-            if (!this.scope.existName(nname)) {
+            if (!scope.existName(nname)) {
                 this.generatedNames.set(name, counter);
                 return nname;
             }
         }
-    }
-}
-
-export namespace UnusedName {
-    export interface Scope {
-        existName(name:string):boolean;
-        canAccessGlobalName(bame:string):boolean;
     }
 }
